+++ src/vonage/client.py
@@ -1,4 +1,5 @@
-from ._internal import ApplicationV2, BasicAuthenticatedServer, _format_date_param
+from .application import Application, BasicAuthenticatedServer
+from ._internal import _format_date_param
 from .errors import *
 from .voice import *
 from .sms import *
@@ -16,7 +17,7 @@ import pytz
 import requests
 import time
 from uuid import uuid4
-import warnings
+import warnings 
 import re
 from deprecated import deprecated
 
@@ -40,7 +41,7 @@ class Client:
     the relevant classes (e.g. `voice.py`, `sms.py`). Change your code to call these classes directly
     as they will be removed in a later release!
 
-    Newer APIs are under namespaces like :attr:`Client.application_v2`.
+    Newer APIs are under namespaces like :attr:`Client.application`.
 
     The credentials you provide when instantiating a Client determine which
     methods can be called. Consult the `Vonage API docs <https://developer.vonage.com/api/>`_ for details of the
@@ -120,7 +121,7 @@ class Client:
             api_key=self.api_key,
             api_secret=self.api_secret,
         )
-        self.application_v2 = ApplicationV2(api_server)
+        self.application = Application(api_server)
 
         self.session = requests.Session()
 
@@ -389,7 +390,7 @@ class Client:
 
     def get_applications(self, params=None, **kwargs):
         warnings.warn(
-            "vonage.Client#get_applications is deprecated (use methods from #application_v2 instead)",
+            "vonage.Client#get_applications is deprecated (use methods from #application instead)",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -397,7 +398,7 @@ class Client:
 
     def get_application(self, application_id):
         warnings.warn(
-            "vonage.Client#get_application is deprecated (use methods from #application_v2 instead)",
+            "vonage.Client#get_application is deprecated (use methods from #application instead)",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -408,7 +409,7 @@ class Client:
 
     def create_application(self, params=None, **kwargs):
         warnings.warn(
-            "vonage.Client#create_application is deprecated (use methods from #application_v2 instead)",
+            "vonage.Client#create_application is deprecated (use methods from #application instead)",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -416,7 +417,7 @@ class Client:
 
     def update_application(self, application_id, params=None, **kwargs):
         warnings.warn(
-            "vonage.Client#update_application is deprecated (use methods from #application_v2 instead)",
+            "vonage.Client#update_application is deprecated (use methods from #application instead)",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -428,7 +429,7 @@ class Client:
 
     def delete_application(self, application_id):
         warnings.warn(
-            "vonage.Client#delete_application is deprecated (use methods from #application_v2 instead)",
+            "vonage.Client#delete_application is deprecated (use methods from #application instead)",
             DeprecationWarning,
             stacklevel=2,
         )
