+++ src/vonage/message_classes.py
@@ -1,0 +1,18 @@
+class MessagesObject(object):
+    pass
+
+class SmsMessage(MessagesObject):
+    def __init__(self):
+        pass
+
+class MmsMessage(MessagesObject):
+    pass
+
+class WhatsAppMessage(MessagesObject):
+    pass
+
+class MessengerMessage(MessagesObject):
+    pass
+
+class ViberMessage(MessagesObject):
+    pass
