+++ tests/test_applications_v2.py
@@ -12,7 +12,7 @@ def test_list_applications(client, dummy_data):
         fixture_path="applications_v2/list_applications.json",
     )
 
-    apps = client.application_v2.list_applications()
+    apps = client.application.list_applications()
     assert_basic_auth()
     assert isinstance(apps, dict)
     assert apps["total_items"] == 30
@@ -27,7 +27,7 @@ def test_get_application(client, dummy_data):
         fixture_path="applications_v2/get_application.json",
     )
 
-    app = client.application_v2.get_application("xx-xx-xx-xx")
+    app = client.application.get_application("xx-xx-xx-xx")
     assert_basic_auth()
     assert isinstance(app, dict)
     assert app["name"] == "My Test Application"
@@ -44,7 +44,7 @@ def test_create_application(client, dummy_data):
 
     params = {"name": "Example App", "type": "voice"}
 
-    app = client.application_v2.create_application(params)
+    app = client.application.create_application(params)
     assert_basic_auth()
     assert isinstance(app, dict)
     assert app["name"] == "My Test Application"
@@ -63,7 +63,7 @@ def test_update_application(client, dummy_data):
 
     params = {"answer_url": "https://example.com/ncco"}
 
-    app = client.application_v2.update_application("xx-xx-xx-xx", params)
+    app = client.application.update_application("xx-xx-xx-xx", params)
     assert_basic_auth()
     assert isinstance(app, dict)
     assert request_user_agent() == dummy_data.user_agent
@@ -81,7 +81,7 @@ def test_delete_application(client, dummy_data):
         status=204,
     )
 
-    assert client.application_v2.delete_application("xx-xx-xx-xx") is None
+    assert client.application.delete_application("xx-xx-xx-xx") is None
     assert_basic_auth()
     assert request_user_agent() == dummy_data.user_agent
 
@@ -94,7 +94,7 @@ def test_authentication_error(client):
         status=401,
     )
     with pytest.raises(vonage.AuthenticationError):
-        client.application_v2.delete_application("xx-xx-xx-xx")
+        client.application.delete_application("xx-xx-xx-xx")
 
 
 @responses.activate
@@ -112,7 +112,7 @@ def test_client_error(client):
         ),
     )
     with pytest.raises(vonage.ClientError) as exc_info:
-        client.application_v2.delete_application("xx-xx-xx-xx")
+        client.application.delete_application("xx-xx-xx-xx")
     assert (
         str(exc_info.value) == "Nope: You really shouldn't have done that (nope_error)"
     )
@@ -127,7 +127,7 @@ def test_client_error_no_decode(client):
         body="{this: isnot_json",
     )
     with pytest.raises(vonage.ClientError) as exc_info:
-        client.application_v2.delete_application("xx-xx-xx-xx")
+        client.application.delete_application("xx-xx-xx-xx")
     assert str(exc_info.value) == "430 response"
 
 
@@ -139,4 +139,4 @@ def test_server_error(client):
         status=500,
     )
     with pytest.raises(vonage.ServerError):
-        client.application_v2.delete_application("xx-xx-xx-xx")
+        client.application.delete_application("xx-xx-xx-xx")
